app/Http/Requests/Manager/Courseディレクトリが肥大化しているための整理(yuki)
