+++ app/Http/Requests/Manager/Course/DeleteRequest.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Manager\Course;
 
-use App\Rules\CourseStatusRule;
 use Illuminate\Foundation\Http\FormRequest;
 
-class CourseUpdateRequest extends FormRequest
+class DeleteRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -33,9 +32,6 @@ class CourseUpdateRequest extends FormRequest
     {
         return [
             'course_id' => ['required', 'integer', 'exists:courses,id,deleted_at,NULL'],
-            'title' => ['required', 'string'],
-            'image' => ['mimes:jpg,png'],
-            'status' => ['required', 'string', new CourseStatusRule],
         ];
     }
 }
