+++ app/Http/Requests/Manager/Course/StatusRequest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Manager\Course;
 
+use App\Rules\CourseStatusRule;
 use Illuminate\Foundation\Http\FormRequest;
 
-class CourseDeleteRequest extends FormRequest
+class StatusRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -16,13 +17,6 @@ class CourseDeleteRequest extends FormRequest
         return true;
     }
 
-    protected function prepareForValidation()
-    {
-        $this->merge([
-            'course_id' => $this->route('course_id'),
-        ]);
-    }
-
     /**
      * Get the validation rules that apply to the request.
      *
@@ -31,7 +25,7 @@ class CourseDeleteRequest extends FormRequest
     public function rules()
     {
         return [
-            'course_id' => ['required', 'integer', 'exists:courses,id,deleted_at,NULL'],
+            'status' => ['required', 'string', new CourseStatusRule],
         ];
     }
 }
