+++ app/Http/Requests/Manager/Course/StoreRequest.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Manager\Course;
 
-use App\Rules\CourseStatusRule;
 use Illuminate\Foundation\Http\FormRequest;
 
-class CoursePutStatusRequest extends FormRequest
+class StoreRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -25,7 +24,8 @@ class CoursePutStatusRequest extends FormRequest
     public function rules()
     {
         return [
-            'status' => ['required', 'string', new CourseStatusRule],
+            'title' => ['required', 'string', 'max:30'],
+            'image' => ['required', 'mimes:jpg,png', 'max:2048'],
         ];
     }
 }
